fix(seo): move buildDatasetJsonLd to _jsonld.ts to avoid invalid Next.js page export
