+++ frontend/app/licitacoes/[setor]/page.tsx
@@ -16,6 +16,7 @@ import { getFreshnessLabel } from "@/lib/seo";
 import { MicroDemo } from "@/components/seo/MicroDemo";
 import { MicroDemoSchema } from "@/components/seo/MicroDemoSchema";
 import StickyTrialCTA from "@/app/components/StickyTrialCTA";
+import { buildDatasetJsonLd } from "./_jsonld";
 
 /**
  * STORY-324 AC5: SSG with ISR 6h for sector landing pages.
@@ -593,85 +594,6 @@ function buildJsonLd(
 }
 
 // SEO-PLAYBOOK P4: Dataset schema for AI Overviews eligibility
-// Always emitted — describes the conceptual dataset (PNCP bids for this sector).
-// Live `total_open` is enriched opportunistically when available.
-export function buildDatasetJsonLd(
-  sector: { name: string; slug: string },
-  stats: SectorStats | null,
-) {
-  const totalOpen = stats?.total_open ?? 0;
-  const hasLiveCount = totalOpen > 0;
-  const description = hasLiveCount
-    ? `Dataset ao vivo com ${totalOpen} licitações públicas abertas de ${sector.name} no Brasil, agregadas do PNCP (Portal Nacional de Contratações Públicas) e atualizadas a cada 6 horas.`
-    : `Dataset ao vivo de licitações públicas de ${sector.name} no Brasil, agregadas do PNCP (Portal Nacional de Contratações Públicas) e atualizadas a cada 6 horas.`;
-
-  const dataset: Record<string, unknown> = {
-    "@context": "https://schema.org",
-    "@type": "Dataset",
-    name: `Licitações de ${sector.name} — Dataset SmartLic`,
-    description,
-    keywords: [
-      `licitações ${sector.name}`,
-      `editais ${sector.name}`,
-      "PNCP",
-      "contratações públicas",
-      "Lei 14.133",
-    ],
-    variableMeasured: [
-      "Total de licitações públicas abertas",
-      "Valor médio por edital",
-      "Órgãos contratantes",
-      "Modalidades de contratação",
-    ],
-    measurementTechnique:
-      "Agregação automatizada via PNCP — Portal Nacional de Contratações Públicas, com deduplicação por content hash e classificação setorial via LLM (GPT-4.1-nano)",
-    temporalCoverage: "2024-01-01/..",
-    spatialCoverage: {
-      "@type": "Place",
-      name: "Brasil",
-      geo: {
-        "@type": "GeoShape",
-        addressCountry: "BR",
-      },
-    },
-    isAccessibleForFree: true,
-    license: "https://creativecommons.org/licenses/by/4.0/",
-    creator: {
-      "@type": "Organization",
-      name: "SmartLic / CONFENGE Avaliacoes e Inteligencia Artificial LTDA",
-      url: "https://smartlic.tech",
-    },
-    publisher: {
-      "@type": "Organization",
-      name: "SmartLic",
-      url: "https://smartlic.tech",
-    },
-    isBasedOn: {
-      "@type": "Dataset",
-      name: "PNCP — Portal Nacional de Contratações Públicas",
-      url: "https://pncp.gov.br",
-      publisher: {
-        "@type": "GovernmentOrganization",
-        name: "Governo Federal do Brasil",
-      },
-    },
-    distribution: [
-      {
-        "@type": "DataDownload",
-        encodingFormat: "application/json",
-        contentUrl: `https://smartlic.tech/v1/sectors/${sector.slug}/stats`,
-      },
-    ],
-    url: `https://smartlic.tech/licitacoes/${sector.slug}`,
-  };
-
-  if (hasLiveCount) {
-    (dataset as Record<string, unknown>).size = `${totalOpen} editais abertos`;
-  }
-
-  return dataset;
-}
-
 // SEO-PLAYBOOK P4: HowTo schema for rich snippets
 function buildHowToJsonLd(sector: { name: string }) {
   return {
